+++ lib/src/vdom/component.dart
@@ -1,14 +1,10 @@
 part of liquid;
 
-abstract class VComponent<C extends Component<T>, T extends html.Element>
-    extends vdom.ElementBase<T> {
+abstract class VComponentBase<C extends Component<T>, T extends html.Element>
+  extends vdom.Node<T> {
   C component = null;
 
-  VComponent(Object key,
-      Map<String, String> attributes,
-      List<String> classes,
-      Map<String, String> styles)
-      : super(key, attributes, classes, styles);
+  VComponentBase(Object key) : super(key);
 
   void create(Context context);
 
@@ -16,7 +12,7 @@ abstract class VComponent<C extends Component<T>, T extends html.Element>
     component.render();
   }
 
-  void update(VComponent<C, T> other, Context context) {
+  void update(VComponentBase<C, T> other, Context context) {
     other.ref = ref;
     other.component = component;
   }
@@ -30,12 +26,57 @@ abstract class VComponent<C extends Component<T>, T extends html.Element>
   }
 
   String toString() => (component == null) ?
+      'VComponentBase[stateless]' : 'VComponentBase[$component]';
+}
+
+abstract class VComponent<C extends Component<T>, T extends html.Element>
+    extends VComponentBase<C, T> {
+  Map<String, String> attributes;
+  List<String> classes;
+  Map<String, String> styles;
+
+  VComponent(Object key, this.attributes, this.classes, this.styles)
+       : super(key);
+
+  void render(Context context) {
+    super.render(context);
+    if (attributes != null) {
+      attributes.forEach((key, value) {
+        ref.setAttribute(key, value);
+      });
+    }
+    if (styles != null) {
+      styles.forEach((key, value) {
+        ref.style.setProperty(key, value);
+      });
+    }
+    if (classes != null) {
+      ref.classes.addAll(classes);
+    }
+  }
+
+  void update(VComponent<C, T> other, Context context) {
+    super.update(other, context);
+    if (attributes != null || other.attributes != null) {
+      vdom.updateMap(attributes, other.attributes, ref.attributes);
+    }
+    if (styles != null || other.styles != null) {
+      vdom.updateStyle(styles, other.styles, ref.style);
+    }
+    if (classes != null || other.classes != null) {
+      vdom.updateSet(classes, other.classes, ref.classes);
+    }
+  }
+
+  String toString() => (component == null) ?
       'VComponent[stateless]' : 'VComponent[$component]';
 }
 
 abstract class VComponentContainer<C extends Component<T>, T extends html.Element>
-    extends VComponent<C, T> {
+    extends VComponent<C, T> with vdom.Container {
   List<vdom.Node> children;
+
+  html.Node get container => component.container;
 
   VComponentContainer(Object key,
       this.children,
@@ -43,55 +84,19 @@ abstract class VComponentContainer<C extends Component<T>, T extends html.Elemen
       List<String> classes,
       Map<String, String> styles)
       : super(key, attributes, classes, styles);
-}
 
-
-/// TODO: rename
-typedef Component VDomInitFunction(Component component,
-                                   vdom.Context context);
-
-/// VDom Node for Components
-class VDomComponent extends vdom.Node {
-  VDomInitFunction _initFunction;
-  Component component = null;
-
-  VDomComponent(Object key, this._initFunction) : super(key) {
-    assert(_initFunction != null);
+  void insertBefore(vdom.Node node, html.Node nextRef, Context context) {
+    component.insertBefore(node, nextRef);
   }
 
-  void create(vdom.Context context) {
-    assert(component == null);
-    assert(ref == null);
-    component = _initFunction(null, context);
-    ref = component.element;
+  void move(vdom.Node node, html.Node nextRef, Context context) {
+    component.move(node, nextRef);
   }
 
-  void render(vdom.Context context) {
-    assert(component != null);
-    assert(ref != null);
-    component.render();
+  void removeChild(vdom.Node node, Context context) {
+    component.removeChild(node);
   }
 
-  void update(VDomComponent other, vdom.Context context) {
-    assert(other != null);
-
-    // transfer component state
-    other.ref = ref;
-    other.component = component;
-    other._initFunction(component, context);
-  }
-
-  void attached() {
-    component.attached();
-  }
-
-  void detached() {
-    component.detached();
-  }
-
-  String toString() {
-    return (component == null)
-        ? 'VDomComponent[stateless]'
-        : 'VDomComponent[$component]';
-  }
+  String toString() => (component == null) ?
+      'VComponentContainer[stateless]' : 'VComponentContainer[$component]';
 }